requirements.txt
Flask>=2.2.0
pycryptodome>=3.17
